chore: extract PoseDetWebcamInner component
